moved BasicAuthHandler out of the BasicAuthCredentials class
