+++ Docker.DotNet.BasicAuth/BasicAuthCredentials.cs
@@ -1,11 +1,6 @@
 ﻿using System;
 using System.Net.Http;
-using System.Net.Http.Headers;
-using System.Runtime.InteropServices;
 using System.Security;
-using System.Text;
-using System.Threading;
-using System.Threading.Tasks;
 
 namespace Docker.DotNet.BasicAuth
 {
@@ -13,47 +8,6 @@ namespace Docker.DotNet.BasicAuth
     {
         private readonly BasicAuthHandler _handler;
         private readonly bool _isTls;
-
-        private class BasicAuthHandler : HttpClientHandler
-        {
-            private readonly SecureString _username;
-            private readonly SecureString _password;
-
-            public BasicAuthHandler(SecureString username, SecureString password)
-            {
-                _username = username;
-                _password = password;
-            }
-
-            protected override Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
-            {
-                request.Headers.Authorization = new AuthenticationHeaderValue("Basic", BuildParameters());
-
-                return base.SendAsync(request, cancellationToken);
-            }
-
-            private string BuildParameters()
-            {
-                string authInfo = string.Format("{0}:{1}", ConvertToUnsecureString(_username),
-                    ConvertToUnsecureString(_password));
-
-                return Convert.ToBase64String(Encoding.Default.GetBytes(authInfo));
-            }
-
-            private static string ConvertToUnsecureString(SecureString secureString)
-            {
-                IntPtr unmanagedString = IntPtr.Zero;
-                try
-                {
-                    unmanagedString = Marshal.SecureStringToGlobalAllocUnicode(secureString);
-                    return Marshal.PtrToStringUni(unmanagedString);
-                }
-                finally
-                {
-                    Marshal.ZeroFreeGlobalAllocUnicode(unmanagedString);
-                }
-            }
-        }
 
         public override HttpMessageHandler Handler
         {
@@ -75,7 +29,7 @@ namespace Docker.DotNet.BasicAuth
                 throw new ArgumentException("password");
             }
 
-            _handler = new BasicAuthHandler(username, password);
+            _handler = CreateHandler(username, password);
 
             _isTls = isTls;
         }
@@ -92,9 +46,14 @@ namespace Docker.DotNet.BasicAuth
                 throw new ArgumentException("password");
             }
 
-            _handler = new BasicAuthHandler(ConvertToSecureString(username), ConvertToSecureString(password));
+            _handler = CreateHandler(ConvertToSecureString(username), ConvertToSecureString(password));
 
             _isTls = isTls;
+        }
+
+        private BasicAuthHandler CreateHandler(SecureString username, SecureString password)
+        {
+            return new BasicAuthHandler(username, password, new HttpClientHandler());
         }
 
         public override bool IsTlsCredentials()
