+++ Docker.DotNet.BasicAuth/Docker.DotNet.BasicAuth.csproj
@@ -48,6 +48,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="BasicAuthCredentials.cs" />
+    <Compile Include="BasicAuthHandler.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
